requirements.txt
discord.py
google-genai
aiomysql
aiohttp
Pillow
duckduckgo-search
python-dotenv
